@@ -15,31 +15,37 @@ import {
   StyledAccount,
 } from './Header.style';
 import logo from '../../media/logo/spiral.svg';
+import { useTranslation } from 'react-i18next';
+import LanguageSelect from '../LanguageSelect';
 
 const AppHeader: React.FC = (props: any) => {
   const { email } = props.auth;
   const { name, type } = props.profile;
   const firebase = useFirebase();
+  const { t } = useTranslation();
 
   const logOutHandler = () => firebase.logout();
 
   return (
     <StyledHeader>
       <StyledLogo>
-        <img src={logo} alt={'site logo'} />
+        <img src={logo} alt={t('header.siteLogo')} />
         <p>volunteer</p>
       </StyledLogo>
+      <Box marginLeft="56px" marginRight="auto">
+        <LanguageSelect />
+      </Box>
       <StyledAccount>
         <Box alignSelf="flex-end">
-          <StyledUserName>{`Name: ${name}`}</StyledUserName>
+          <StyledUserName>{`${t('header.name')}: ${name}`}</StyledUserName>
           <StyledInfoContainer>
-            <p>{`Role: ${TITLE_USER_TYPE_MAP[type]}`}</p>
-            <p>{`Email: ${email}`}</p>
+            <p>{`${t('header.role')}: ${TITLE_USER_TYPE_MAP[type]}`}</p>
+            <p>{`${t('header.email')}: ${email}`}</p>
           </StyledInfoContainer>
         </Box>
         <Box alignSelf="flex-end">
           <IconButton onClick={logOutHandler}>
-            <ExitToAppIcon>logOut</ExitToAppIcon>
+            <ExitToAppIcon>{t('header.logOut')}</ExitToAppIcon>
           </IconButton>
         </Box>
       </StyledAccount>
