@@ -1,6 +1,7 @@
 import React from 'react';
-import loadingIcon from '../../media/icons/loading.svg';
+import { useTranslation } from 'react-i18next';
 
+import loadingIcon from '../../media/icons/loading.svg';
 import { StyledLoadingContainer, LoadingIcon, LoadingText } from './Loading,style';
 
 interface LoadingProps {
@@ -9,10 +10,12 @@ interface LoadingProps {
 }
 
 export const Loading: React.FC<LoadingProps> = ({ className = '', showText = false }) => {
+  const { t } = useTranslation();
+
   return (
     <StyledLoadingContainer className={className}>
       <LoadingIcon src={String(loadingIcon)} alt="loading" />
-      {showText ? <LoadingText>Loading...</LoadingText> : null}
+      {showText ? <LoadingText>{t('loading')}...</LoadingText> : null}
     </StyledLoadingContainer>
   );
 };
