@@ -14,6 +14,7 @@ import { filter, map, switchMap, withLatestFrom } from 'rxjs/operators';
 import { IAppState } from '../interfaces/state';
 import { notify } from './notification';
 import selectors from '../selectors';
+import i18n from '../i18n';
 
 const firebasePrefix: string = '@@reactReduxFirebase';
 
@@ -45,7 +46,7 @@ const startFetchUsers = (action$: ActionsObservable<Action>, state$: StateObserv
     ofType(`${firebasePrefix}/SET_PROFILE`),
     withLatestFrom(state$),
     filter(([action, state$]) => selectors.user.getEmployeeType(state$) !== EmployeeType.Volunteer),
-    map(() => fetchDataRequested())
+    map(() => fetchDataRequested()),
   );
 
 const fetchUsers = (action$: ActionsObservable<Action>, state$: StateObservable<IAppState>) =>
@@ -56,16 +57,14 @@ const fetchUsers = (action$: ActionsObservable<Action>, state$: StateObservable<
     switchMap(([action, state]) =>
       getQuery(state.firebase)
         .then((data) => fetchDataFulfilled(data.val() || {}))
-        .catch((error) => fetchDataFailed({ error: error.code }))
-    )
+        .catch((error) => fetchDataFailed({ error: error.code })),
+    ),
   );
 
 const onFetchDataError = (action$: ActionsObservable<getUsersFail>) =>
   action$.pipe(
     filter(isOfType(GET_USERS_FAIL)),
-    map((action) =>
-      notify(`Unexpected error while loading usersData. Code: ${action.payload.error}`)
-    )
+    map((action) => notify(i18n.t('notifications.usersDataError', { code: action.payload.error }))),
   );
 
 export default [startFetchUsers, fetchUsers, onFetchDataError];
