@@ -1,18 +1,28 @@
-import { ActivityStatus, ActivityType } from 'common/index';
+import { ActivityStatus, ActivityType, ActivityPriority } from 'common/index';
 
 export const TITLE_STATUS_MAP = {
-  [ActivityStatus.New]: 'Created',
-  [ActivityStatus.ReadyForAssignment]: 'Backlog',
-  [ActivityStatus.Assigned]: 'Assigned',
-  [ActivityStatus.InProgress]: 'In Progress',
-  [ActivityStatus.Canceled]: 'Canceled',
-  [ActivityStatus.Done]: 'Done',
-  [ActivityStatus.Archived]: 'Archived',
+  [ActivityStatus.New]: 'statuses.created',
+  [ActivityStatus.ReadyForAssignment]: 'statuses.backlog',
+  [ActivityStatus.Assigned]: 'statuses.assigned',
+  [ActivityStatus.InProgress]: 'statuses.inProgress',
+  [ActivityStatus.Canceled]: 'statuses.canceled',
+  [ActivityStatus.Done]: 'statuses.done',
+  [ActivityStatus.Archived]: 'statuses.archived',
 };
 
 export const TITLE_TYPE_MAP = {
-  [ActivityType.Delivery]: 'Delivery',
-  [ActivityType.HomeCare]: 'Home care',
-  [ActivityType.Other]: 'Other',
-  [ActivityType.Shopping]: 'Shopping',
+  [ActivityType.Other]: 'types.other',
+  [ActivityType.shoppingDelivery]: 'types.shoppingDelivery',
+  [ActivityType.HomeCare]: 'types.homeCare',
+  [ActivityType.eco]: 'types.eco',
+  [ActivityType.hospital]: 'types.hospital',
+  [ActivityType.disabledPeople]: 'types.disabledPeople',
+  [ActivityType.orphanage]: 'types.orphanage',
+  [ActivityType.events]: 'types.events',
+};
+
+export const PRIORITY_TYPE_MAP = {
+  [ActivityPriority.Low]: 'priority.low',
+  [ActivityPriority.Middle]: 'priority.middle',
+  [ActivityPriority.High]: 'priority.high',
 };
